+++ src/object/hit.rs
@@ -3,13 +3,13 @@ use std::{ops::Range, rc::Rc};
 use crate::geometry::{Point, Ray, Vec3};
 
 pub struct HitRecord {
-    pub point: Point<f64>,
-    pub normal: Vec3<f64>,
-    pub t: f64,
+    pub point: Point<f32>,
+    pub normal: Vec3<f32>,
+    pub t: f32,
 }
 
 pub trait Hit {
-    fn hit(&self, r: Ray, t_range: Range<f64>) -> Option<HitRecord>;
+    fn hit(&self, r: Ray, t_range: Range<f32>) -> Option<HitRecord>;
 }
 
 /// Stores a list of references to Hit objects
@@ -31,7 +31,7 @@ impl HittableList {
 
 impl Hit for HittableList {
     /// Returns the closest hit from hitting all elements in the list
-    fn hit(&self, r: Ray, t_range: Range<f64>) -> Option<HitRecord> {
+    fn hit(&self, r: Ray, t_range: Range<f32>) -> Option<HitRecord> {
         let mut closest_t = t_range.end;
 
         self.objects
@@ -40,7 +40,7 @@ impl Hit for HittableList {
                 let hit = x.hit(r, t_range.start..closest_t);
 
                 if let Some(ref record) = hit {
-                    closest_t = f64::min(closest_t, record.t);
+                    closest_t = f32::min(closest_t, record.t);
                 }
 
                 hit
